Add SQS to mirrord operator status reporting.
