+++ mirrord/cli/src/operator/status.rs
@@ -1,0 +1,225 @@
+use std::{path::PathBuf, time::Duration};
+
+use futures::{StreamExt, TryStreamExt};
+use mirrord_analytics::NullReporter;
+use mirrord_config::{
+    config::{ConfigContext, MirrordConfig},
+    LayerConfig, LayerFileConfig,
+};
+use mirrord_kube::api::kubernetes::seeker::KubeResourceSeeker;
+use mirrord_operator::{
+    client::{error::OperatorApiError, NoClientCert, OperatorApi},
+    crd::{MirrordOperatorSpec, MirrordWorkloadQueueRegistry},
+    types::LicenseInfoOwned,
+};
+use mirrord_progress::{Progress, ProgressTracker};
+use prettytable::{row, Table};
+use tracing::Level;
+
+use crate::{error::CliError, util::remove_proxy_env, CliResult};
+
+/// Handles the `mirrord operator status` command.
+pub(super) struct StatusCommandHandler {
+    /// Api to talk with session routes in the operator.
+    operator_api: OperatorApi<NoClientCert>,
+}
+
+impl StatusCommandHandler {
+    #[tracing::instrument(level = Level::TRACE, err)]
+    pub(super) async fn new(config_file: Option<PathBuf>) -> CliResult<Self> {
+        let mut progress = ProgressTracker::from_env("Operator Status");
+
+        let layer_config = if let Some(config) = config_file {
+            let mut cfg_context = ConfigContext::default();
+            LayerFileConfig::from_path(config)?.generate_config(&mut cfg_context)?
+        } else {
+            LayerConfig::from_env()?
+        };
+
+        if !layer_config.use_proxy {
+            remove_proxy_env();
+        }
+
+        let mut status_progress = progress.subtask("fetching status");
+        let api = OperatorApi::try_new(&layer_config, &mut NullReporter::default())
+            .await
+            .inspect_err(|_| {
+                status_progress.failure(Some("failed to get status"));
+            })?
+            .ok_or(CliError::OperatorNotInstalled)?;
+
+        status_progress.success(Some("fetched status"));
+        progress.success(None);
+
+        Ok(Self { operator_api: api })
+    }
+
+    #[tracing::instrument(level = Level::TRACE, skip(self), ret, err)]
+    pub(super) async fn handle(self) -> CliResult<()> {
+        use std::future::ready;
+
+        let Self { operator_api: api } = self;
+
+        let MirrordOperatorSpec {
+            operator_version,
+            default_namespace,
+            license:
+                LicenseInfoOwned {
+                    name,
+                    organization,
+                    expire_at,
+                    ..
+                },
+            ..
+        } = api.operator().spec.clone();
+
+        let expire_at = expire_at.format("%e-%b-%Y");
+
+        println!(
+            r#"
+Operator version: {operator_version}
+Operator default namespace: {default_namespace}
+Operator License
+    name: {name}
+    organization: {organization}
+    expire at: {expire_at}
+"#
+        );
+
+        let status = api
+            .operator()
+            .status
+            .as_ref()
+            .ok_or(CliError::OperatorStatusNotFound)?;
+
+        if let Some(copy_targets) = status.copy_targets.as_ref() {
+            if copy_targets.is_empty() {
+                println!("No active copy targets.");
+            } else {
+                println!("Active Copy Targets:");
+                let mut copy_targets_table = Table::new();
+
+                copy_targets_table.add_row(row![
+                    "Original Target",
+                    "Namespace",
+                    "Copy Pod Name",
+                    "Scale Down?"
+                ]);
+
+                for (pod_name, copy_target_resource) in copy_targets {
+                    copy_targets_table.add_row(row![
+                        copy_target_resource.spec.target.to_string(),
+                        copy_target_resource
+                            .metadata
+                            .namespace
+                            .as_deref()
+                            .unwrap_or_default(),
+                        pod_name,
+                        if copy_target_resource.spec.scale_down {
+                            "*"
+                        } else {
+                            ""
+                        },
+                    ]);
+                }
+
+                copy_targets_table.printstd();
+            }
+            println!();
+        }
+
+        if let Some(statistics) = status.statistics.as_ref() {
+            println!("Operator Daily Users: {}", statistics.dau);
+            println!("Operator Monthly Users: {}", statistics.mau);
+        }
+
+        let mut sessions = Table::new();
+
+        sessions.add_row(row![
+            "Session ID",
+            "Target",
+            "Namespace",
+            "User",
+            "Ports",
+            "Session Duration"
+        ]);
+
+        for session in &status.sessions {
+            let locked_ports = session
+                .locked_ports
+                .as_deref()
+                .map(|ports| {
+                    ports
+                        .iter()
+                        .map(|(port, type_, filter)| {
+                            format!(
+                                "Port: {port}, Type: {type_}{}",
+                                filter
+                                    .as_ref()
+                                    .map(|f| format!(", Filter: {}", f))
+                                    .unwrap_or_default()
+                            )
+                        })
+                        .collect::<Vec<_>>()
+                        .join("\n")
+                })
+                .unwrap_or_default();
+
+            sessions.add_row(row![
+                session.id.as_deref().unwrap_or_default(),
+                &session.target,
+                session.namespace.as_deref().unwrap_or("N/A"),
+                &session.user,
+                locked_ports,
+                humantime::format_duration(Duration::from_secs(session.duration_secs)),
+            ]);
+        }
+
+        sessions.printstd();
+        println!();
+
+        let mut reporter = NullReporter::default();
+        let api = api.prepare_client_cert(&mut reporter).await;
+        api.inspect_cert_error(
+            |error| tracing::error!(%error, "failed to prepare client certificate"),
+        );
+
+        let seeker = KubeResourceSeeker {
+            client: api.client(),
+            namespace: Some(default_namespace.as_str()),
+        };
+
+        let mut sqs = Table::new();
+        sqs.add_row(row!["Queue Name", "Consumer"]);
+
+        for sqs_row in seeker
+            .list_resource::<MirrordWorkloadQueueRegistry>(None)
+            .filter(|response| ready(response.is_ok()))
+            .try_collect::<Vec<_>>()
+            .await
+            .map_err(OperatorApiError::from)?
+            .into_iter()
+            .filter_map(|sqs| {
+                let status = sqs.status?;
+                let details = status.sqs_details?;
+                let consumer = sqs.spec.consumer;
+                let names = details
+                    .output_queue_names()
+                    .into_iter()
+                    .map(|name| row![name, &consumer]);
+
+                Some(names.collect::<Vec<_>>())
+            })
+            .flatten()
+        {
+            sqs.add_row(sqs_row);
+        }
+
+        if sqs.len() > 1 {
+            println!("Active SQS Queues:");
+            sqs.printstd();
+        }
+
+        Ok(())
+    }
+}
